Fixes Issue #39 - Create a test to check One and Done Home Page
